check trùng danh mục thử nghiệm màn tích điểm

+++ src/main/java/com/example/datn404shoes/entity/TichDiem.java
@@ -24,6 +24,6 @@ public class TichDiem {
 
 
     @Column(name = "diem")
-    private Integer diem;
+    private Float diem;
 
 }
